Predictive Timer Logic V2
This change creates the second generation prediction timer logic for
RCP.  Namely this logic bases its guesses using liner interpolation on
the distance between the two closest points points from the fast lap
and those points laptime information.

It works well when GPS accuracy is good as it relies solely on GPS.
Should the accuracy drop off the algorithm will screw up.  Of course
this is not the algorithms fault as its just getting bad/inaccurate
data.  In other words, accurate GPS readings are a must for this to
work accurately.  Also a consistent flow of readings is also necessary
as these are stored and used if the lap was the fast lap.