Add Automatic Track Detection Logic
Adds primitive automatic track support logic to RaceCapture.  This
support gets enabled when the track start/finish line values are
configured to 0.  If the start/finish line values are configured to
anything else then this feature will be disabled.

No testing has been written as of yet (next up) and there are not many
tracks that RC currently has programmed in.  Both are issues to be
solved in upcoming patches.