Remove Eclipse related files and ignore them going forward
Removes the Eclipse files from the project because developers
can and will use different versions of eclipse and we don't want
that screwing with commits (imagine the conflict resolution on
those files :o).

This commit also adds the remove file types to the .gitignore
file so that we ignore them going forward.