Add millis since epoch and re-arrange when onLocationUpdate is called
Adds milliseconds since epoch.  This is quite useful for having a
timing mechanism that always counts up and doesn't roll over unlike
the currently implemented seconds since midnight.

This change also puts the call to onLocationUpdated in the GGA
stanza _before_ the GGA data is parsed.  This is done because GGA
sentences are always first in a NMEA paragraph from our GPS mouse.
By doing this we ensure we read the entire previous paragraph before
we process any of the data for that paragraph.  This allows us to use
all the data whenever we make an update.