Convert to using seconds since first fix for predictive timer
We used to use seconds since midnight for the predictive timer.
There was a major flaw in this though, it would reset whenever
UTC time would roll over to the next day (4pm PDT).  This would
break a persons lap and would screw up the timer until the box
was rebooted.  With this fix we are using absolute time since
the box first got its GPS position.  This frees us from worrying
about the roll over.