Add LUA getTickCount and getTicksPerSecond methods
Until we get a timer we may need a form of a timer in the LUA script.
The best we can do until said timer exists is to use the TickCount
as provided by the FreeRTOS scheduler.  When both of these methods
are combined a user can get a _rough_ idea of how much time has passed.

Currently ticks happen at 200Hz, so a 5ms resolution is pretty good
for LUA scripting.  If they need more we need to implement at hardware
timer, but we already know that.