Get rid of Channel from our firmware and deal with the fallout
Removes the concept of Channel from the firmware.  This change has
large impact on API as it moves the channel definition from being
hardcoded in the firmware to being defined by the RCP application
when it programs the channels.

Still lots of work to do with cleaning up and redicuing size.  This
change is in so that the API get changed to the stable version for
release of MK2.