Cleanup the onTickHandler
This removes the onTickHandler file and consolidates the code into
a source file.  Did this beacuse I spent way to much time trying
to chase down how something worked only to be stumped by an extern
that was hiding the use of a variable.  Externs suck and should not
be used for this exact reason.  So this is why this change is going in.