Refactor code to gain about 300 bytes of space back
This change is part of a larger effort to standardize
our code base a bit.  Part of the motivation is to ensure
that our code base is small and takes up as little room as
possible.  This change improves upon our code by removing
redundant code that was eating space.

This change also applies formatting standards to the code
and adds proper headers so we look like a professional group
or something :).