Add the elapsed lap time channel
Adds support for an elapsed time channel.  Elapsed lap time is the
amount of time that has elapsed since the lap began.  Like many other
methods internally it works in millisecond resolution but has a
conversion to minutes per the old API standard.

This feature also has its own unit test.  I have gotten quite sick of
maintaining that beast of a unit test that is our sector test and
figured this was a good time to break the habbit of feeding that
monster.  Thus this change has its own isolated unit tests under
lap_stats in the test directory.