Add in GeoTrigger support to replace start/finish cool off timer
Before this change we used a timer in the start/finish area.  The
intention was to de-bounce a start/finish cross event with our GPS
system.  This is needed because we use GeoCircles for our boundary
detection.

The assumption was that drivers would not get stuck in the
start/finish zone for more than 10 seconds as, well, its a race track
right.  Yeah turns out that in some races that does in fact happen and
we have had at least one customer write in complaining about this
issue.

So the timer solution was insufficient.  Hence GeoTriggers.  The
concept of these is based of a Schmitt Trigger in EE.  You have to get
X distance from the Start/Finish area before we allow an event to be
registered.

By doing this we effectively debounce start/finish without the use of
a timer and can now support a driver literally parking in start/finish
for an hour and not having us register a lap inaccurately.