Fix data bugs in distance calculation
Since I added the fixes to the stage logic I introduced a regression
where we were double adding the distance at start.  This would cause
a minor skew in the distances compared to our previous versions of
Race Capture.

Fix was to move all data processing prior to start/finish logic and
events and to place all checks of whether or not to be recording into
the objects and methods themselves.

I also exposed the method that tells us if we are in the middle of
recording a lap.  Its not currently needed but probably not a bad
thing to have exposed (like for say LUA).