Redo the variable and method names to our style.  Add TEST_STATIC
Change does two important things.  First it redoes the entire commit
in the RCP style (Linux based with tweaks).  Second it defines and
makes use of TEST_STATIC.  This is a macro that gets defined to nothing
when we are compiling test code, but is defined as 'static' when we
are compiling firmware code.  This is so that we can test methods that
are normally not accessible while keeping good isolation in our firmware
so the linker can better optimize.