Fix up casting so that we can move to a 64bit compiler
This is the last change that was needed so that we could complete
the transition to a system that would build unit tests on both
OSX and Linux.

This change fixes the casting in the date_time_test so that it will
build in x86_64 architecture.  This makes the builds faster and it
opens up the ability to switch to the better clang frontend of
llvm.