Rewrite Filewriter to avoid empty logfiles
The old Filewriter code had a bug where it would open a new log file
and immediately close.  This was due to a logic error in a very messy
piece of code.  Rather than just simply fix it (and leave room for other
errors down the line) I redid the method and broke its sub methods out
into smaller and much more readable methods.  Now its much easier to follow
the state and behavior.

Need to add unit tests to ensure this is all chipper.  Also need to consider
how we will handle cases where we will need to roll over to a new file if
the old one hits the 2G file limit on FAT32 systems.