Improve logger writing capabilities and add testing
This change improves the logger writing capabilities by changing
the logic to wisely be able to restart logging the stream should
the SD card get ejected in the middle of the writing process and
re-inserted.

It fixes several errors including previously not correctly reporting
file writes, not seeking to the end of the file once reopened, and
a couple of other small gotchas.

It also improves on the original design by adding testing and using
f_write to avoid unnecessary memory copying.