Change SDcard f_sync rate to 1000ms to reduce chances of data loss
The original value of 5000ms was too long because in testing the
removal and re-insertion of the SD card I found that we would often
loose chunks of data due to this large gap between syncs.  However
when the value was set to 1 second this mostly went away.

By doing this we will sync 5 times as often, but with less data and
will likely have less loss should someone screw things up.