Create a Vagrant image for firmware builds
This image will allow developers on a non-linux environment to compile
the firmware for RCP locally on their machine through the use of Vagrant.
Simply run 'vagrant up' followed by 'vagrant ssh' to get in.  cd to /vagrant
and run the standard make commands.

TODO: Still having an issue with the setup of the f4_bootloader and its
python dependencies.  Subsequent patches will hopefully fix this.