Support validation of LogMessages and fix several bugs
The main intent of this fix is to validate LogMessages.  Specifically
this means that we validate that the sample buffer that a LogMessage
points to has not been over-written/re-used since the LogMessage was
queued.  To achieve this we timestamp both the LogMessage and the
sample buffer.  The LogMessage is deeply copied on the queue while the
sample buffer isn't.  If when the LogMessage is read we find that the
times are different, then we know the LogMessage is pointing to a
buffer that has been overwritten since the LogMessage was queued and
thus we skip it as it is invalid.  This will in turn prevent us from
recording empty log lines as we are (mostly) ensured that the sample
buffer will not change out from underneath us when reading it.  If
that does begin to happen (as is possible since this is a pre-emptive
system) then we must add a lock to these sample structures to prevent
the behavior.

This change also fixes some very bad behavior where we were popping
LogMessages off of the queue into an undefined space.  This was
occuring because we were allocating pointers to point to areas where
the xQueueReceive should copy its contents to instead of physically
allocating the space.  I'm honestly not quite sure how this never
caused any major issues, but somehow we seemed to have gotten away
with it.  Now the space is clearly defined on the stack.