Implement the bearing channel
Implements the channel that gives you the GPS bearing.  This is
effectively known in navigation as the track but since that word
is already in use its better to use bearing as its close enough.
The bearing is the GPS bearing from previous point to the current
point.  Its represented in degrees where 0 is north, 90 is east,
180 is south and 270 is West.