Fix queue object sizes in the LoggerMessage queues
There was a bug where we were copying objects the size of a
ChannelSample pointer in our Queues instead of one the size of a
LoggerMessage.  This meant that we were not copying enough data over
and thus were reading bogus data when we were doing any operation with
the queue.

The patch corrects this by creating common methods for createing
logger message queues along with methods to send, receive and validate
LoggerMessages with attached samples.