Replace the file_buffer with our ring buffer
We were using a file_buffer that unfortunately had a bug where
improper updating of index values would lead to data loss.  To
solve this I could have fixed that but instead opted to use our
ring_buff code as that is proven to work.

NOTE: There is still a bug... this is only up for early review
purposes at the moment.