Add _RCP_BASE_FILE_ Macro for defining filenames
GCC has the _FILE_ macro, but this foolishly put the full length
of the file path into the file instead of the shorter basename
version of the file.

This patch addresses that by creating a new definition that only
puts the basename of the file into the macro.  This makes strings
shorter and makes life easier for us.