Handle NULL strings in put_string method
It seems that MK1 likes to pass NULL strings like a total jerk.
This was causing it to lock up pre the 2.8.4 release.  Wonderful.
This change addresses that issue by having the put_string method
check if the string is NULL, and if so ignoring it (as its a no-op
anyways).

This patch also adds the unit tests to ensure that it works.