Add another NULL check in the append method, bump up stack size
Adds the NULL check to the append method because that method will
segfault on a NULL string prior to calling into the put_string method.

Also bumps up the stack size a bit because we allocate some larger
chunks for copying and go deeper into the stack with the use of the
ring buffer.