Remove extranious #if's in the code
There were many of them that really didn't need to be there.
Like around the header files (which makes things really ugly)
and in the header files (although I am open to re-considering
this).

The goal here is to minimize the #ifs so that we can keep our
code as clean as we can until we can make things more modular
and remove the #if's entirely.