Get watchdogs operational on stm32f3 processor
This change had the fix needed to make the watchdog operational
on the stm32f3 processor.  Tested using blinky task on personal
laptop dev and it worked fine for both positive and negative test.

Currently the flags I get back from the RCC registers seem to be
wrong as both the power on reset and the watchdog reset flags are
set when I read them from the RCC.  Don't know why but those are
not critical for operation.