Cleanup the LED code
Initially I thought that the multiple calls to the initiaize GPIO code
were causing previous initilizations to become undone.  After reading
the code I have since learned that this is not the case.

But in the process of cleanup, I did find a minor bug in how we
maintain state, so I fixed that and now the toggle functions will work
as expected.