Turn off the LUA memory optimizations
The memory opts are supposed to reduce memory usage while preseving
basic functionality of the various libraries that we implement.
However this was not the case as LUA was crashing whenever someone
would try to access these libraries.  Thus this change works around
the issue simply by disabling the memory optimizations.  The tradeoff
is ~1KB more usage of RAM compared to before.