Reduce the MK1 LUA script size to 8K from 10K
Our practices behind flashing data sections (LUA, tracks, LoggerConfig)
is a bit arcane in that we allocate the size of the sector to flash
in RAM prior to flashing it.  This means that if we don't have enough
RAM available, it will fail :(.  This was happening on MK1 and is what
was causing many customer issues.

By reducing the size of the LUA script section we use, we regain the
functionality, but we are also burying the issue until we run out of
RAM again.  Ideally the best solution would be not to allocate the
RAM and to instead write the data as we receive it.  However this comes
with its own complications (like how to deal with interrupted/corrupted
transfers) and is not something worth solving right now.