Stop LUA before we reprogram it
We unfortunately program our LUA script in a resource heavy way such
that we tend to run out of available HEAP before we are able to load
the script into flash.

To solve this problem, I have re-tooled our LUA script setup so that
we will unload LUA _before_ we start programming the LUA script.  This
makes sense becuase it will free up LUA resources when they are not
needed (since we are programming and going to restart LUA anyways).
This allows us to keep the 10K script size for MK1.  Wheee fun with
little RAM systems.