Add upper limits for LUA memory usage
If a LUA script is hogging memmory, then it would be able to
destabilize the entire system.  This change adds an upper memory limit
to the LUA resource usage.  By doing this LUA can no longer crash the
system with a memory hungry app.  Instead it will simply cause the LUA
task to reset itself if 3 consecutive memory allocations fail.