Add LUA garbage collection parameters to the lua_initialize
These values allow us to control how our garbage behavior will act
on Race Capture LUA.  This allows us to tune the GC to be more
agressive when we find ourselves working in low memory situations
(as is the case for MK1).  Setting them to 0 means that the default
values as provided by the LUA system are used.