Fix MK1 Crash Bug, Cleanup Large Flashing Routines
Flashing the tracks DB was causing MK1 to crash because it
was allocating a large chunk of RAM to do it.  We need to
disable LUA before we do this, else we will die due to RAM
starvation.  This fix does that and cleans up both of the
large flashing methods because they really could use it.  Now
both methods are much cleaner, up to CONTRIBUTING spec, and
generally make me happy.