Increase size of txrx queue for telem on MK1 and cleanup
This change primarily cleans up the code on MK1 so that we can
easily set different queue sizes for the character queues used
to recieve data on the USART devices.  By doing this, I can set
a larger queue for the BT device, which has been overflowing in
high CPU scenarios.

BT queue is now a symmetric 512/512 for tx/rx queues respectively
and GPS queue is 512/128 (since we don't send much of anything to
the GPS unit).  I increased the GPS unit queue in case we fall a
bit behind in processing its data just as a safety.