Improve the USART code for MK1
This is the more complete fix that was kicked out of 2.8.5 in favor
of a smaller patch b/c of the proximity of the release date.  This
patch cleans up a lot of the UART code and makes it better by
reusing code paths and eliminating externs and duplication that was
found in the previous implementation.