Ensure that a user can never fully disable telemetry
RCP is rather useless if a user disables telemetry completely.
Since we have channels that should always be logging (interval and
UTC) then we should have their default minimum (1Hz) be the default
minimum for them.  This way telemetry is always gauranteed to work
in some form, even if a user foolishly disables all channels.