Add uniqueness to BT name
Upon factory reset, the Bluetooth configu will get a unique suffix
added to the name so that people can more easily differentiate their
BT units from others in the field.  Instead of saying "RaceCapturePro"
as the BT name, it will now read "RaceCapturePro-XXXX" where XXXX
denotes the last 4 characters in the CPU serial number.

This change also changes the default passcode to "0000".  Reason being
is that many units consider this to be the default passcode and often
will not promot the user for password input, making the experience
better for the customer.