Increase strictness of compilation for MK2
This change adds more checks into the compilation process for MK2
so as to ensure better code.  I would have simply used -Wextra but
there are some GCC bugs RE initializers that are not worth
fighting/fixing.  So instead I peacemeal added in the -W* flags
as needed and not buggy.  Would have added -Wunused-parameter but
this needs a lot of love in the code to be able to support this.
Perhaps in a later change.  Not a big deal honestly.