Migrate cellular code to serial methods.  Fix printk methods
Migrates the cellular code to not use the wrapped methods that
used to be in cellular and instead uses only serial methods.

Also fixes up printk by fixing a small bug and adding support
for printing a single character at a time.  This is needed so
we can process the serial messages and ensure that they get
printed to the log correctly.