Fix up start logic on lapstats.  Deal with cascading changes too
This change originally was intended to fix up some possibly buggy
code in the way we process the start point in circuit & stage.  We
lacked the check for the start point and were instead relying on the
debouncing trigger to act as the start point check.  This was a bug
and could cause undefined behavior.

In the process of fixing this, changes to method signatures lead to
all sorts of fallout in how methods are called.  The result is simpler
methods with slight dependence on state of track and defined radius
for geo_circles.