Improve unit tests on the lap_stats logic
Before this commit we had very spotty unit tests on our lap_stats
logic.  This means that it was very easy for little bugs to slip
through un-noticed until users were at track.  This is how #359
likely came to be.

This commit drastically improves the unit tests for this code by
explicitly testing all methods that will change internal state
of the unit.  It replaces both the currentLap and elapsedLapTime
tests as these tests are more comprehensive and are grouped based
on the methods in lap_stats.

Issues: #359