Fix terminal pass-through mode on firmware
This terminal pass-through mode regressed in 2.8.5 (caused a watchdog
every time it was run).  Cause was due to the priority shift of the
tasks.  Since this caused a busy while loop in a high priority task,
lower priority tasks (such as the logger which resets the watchdog)
were starved of resources and thus the issue arose.

This patch addresses the issue by draining both tx and rx buffers on
every pass and then delays for a bit before running again.  This
prevents this busy while from starving the other tasks.

Issues: #352