Add timeout support to the CAN Tx method on MK2
Previously ignored the timeout value and always assumed that the
message would send in time.  Now we will pay attention to the value
and will cancel the message if it has not sent in the specified
timeout.

A timeout value of 0 means we don't care whether it sent or not and
will always return success.

Issue: #404