Add DEBUG flag to MK2 Makefile to make debugging easier.
Adds a DEBUG option to the Makefile so that when you want to enable
features like stackoverflow detection simply re-build with that flag
enabled and it will have it.

This patch also tries to improve the flashing mechanism on the front
of RCP in the event of some form of fault.  As it was lights were
blinking so fast that nothing useful was being seen.