Move IMU mapping changes into HAL layer for MK2.  Fix fallout.
This change implements the conformity to the SAE J670E standard in a
better manner by adjusting the IMU mappings to RCP default coordinate
mappings in the HAL layer instead of in the RCP common layer.

It also addresses the fallout from implmenting enum types instead of
using raw ints for channel values.  It also removes some unnecessary
methods, puts the enum values in the correct places, and fixes the
broken calibration methods on MK2 (MK1 values are not what MK2 values
are, this was never addressed before).