Remove magic numbers.  Apply feedback.  Make jsmn_* methods proper
Moves all relevant magic numbers into #defines at the top of the
file to ensure that no voodoo happens that can't easily be tracked
down.

Fixes up the new jsmn methods so that they are usable from all
locations and are easy to grok.

And generally applies Brent's feedback and suggestions into the
code.