Make the RPM input timers reset themselves when no signal detected
Previously the logic that would detect a buffer overflow in the timer
was disabled because of unexpected interrupts occuring.  These were
caused by the UpdateRequestSource flag not being set to local events
only.

This patch sets that value for the timers 0 & 1 and enables the interrupt
code that allows for clearing of the value when overflow hits (ie
when we lose signal from the engine).  Timer 3 still has emotional issues.

Part of the fix for issue #418