Get the reset logic working on logical timer 2 (Physical 2)
Due to the hardware wiring bug we have no way to have the slave
controller reset the counter when an interrupt comes in.  Thus
the only way we can reset the counter is... not.  This patch
adds code so that we account for roll over, count the ticks per
overflow period, and wisely choose when to reset the output vs
not.  This way we emulate the effects of the slave reset without
having one.