Add in software RPM filtering
Adds in a hard limited software based filter that clips the RPM under
20,000 RPM.  It achieves this by adding a "quiet period" after a
rising edge is detected.  When an interrupt occurs during the quiet
period, the value capture register is stored (contains timer ticks)
but the period/duty_cycle value is not updated.  Outside of the quiet
period, when an interrupt occurs, the sum of the stored ticks is added
to the current value capture register value.  This is then used to
calculate the period.

Tested this by creating a Lua script that creates a dirty signal using
a PWM out port and bit-banging.  When the dirty signal is present, the
RPM value is a bit unstable, but well more than an order of magnitude
improved.  When tested on the current release, the script resulted in
RPM spikes as high as 47,000 RPM.  So I would say it is effective.

This patch is designed to go into 2.8.8 as there are no new parameters
that have been added.  I set the upper limit to support vehicles whose
redline is <= 20,000 RPM.  In 2.9.0, I will add the ability to
dynamically configure this limit so it can be applied enabled,
disabled, or adjusted depending on user's needs.  Until then... it is
what it is.