Prevent unnecessary work when create_virtual_channel gets abused
If a user beats on this call in their Lua script (ie. call it when
a channel of that name has already been created) it will cause us
to constantly invoke the `configChanged()` method, which in turn
causes us to re-allocate all of our buffers used for samples.  This
is disruptive and probably a bad idea, so this prevents that from
happening by bailing out of the call before we get to that point.