Improve the setTickRate setter method
This patch improves the `setTickRate` method by enforcing that it is
only passed one argument, and that that argument is a number.  It also
now returns the interval time in ms that the undelying system will
wait between calls if the method was successful.  If the method was
not successful (invalid rate) it returns an error with a description.