Create validation for ChannelConfig.  Use it on virtual channel creation.
Previously we never had any validation of ChannelConfig in a centralized
location.  This meant if there was any... it was scattered around the
code base and was potentially inconsistent.  Since we now had a need for
some due to issue #424, I decided to go all in and do it properly.

This patch first creates a channel_config file to house all methods
associated with the channel_config structure.  Then it adds methods
to create valid default values of ChannelConfig as well as validators
to ensure that what is in there is valid.  The validators return enum
values to allow for fine grained error detection.  With this in place,
I added thest checks to the addChannel Lua method so that no user is
now able to create a virtual channel with an invalid configuration.

I also backported an improvement from 2.9.0 that allows us to send errors
back to the Lua runtime instead of returning silly values that may silently
break the Lua script.  This will be much more useful in debugging since they
actively tell the user what went wrong (and even give line numbers).

It also adds tests for the new ChannelConfig logic.
Issue: #424