Change serial_buffer_rx to return a pointer to the msg just read
Previously it was returning the number of characters written.  This is
not a useful as returning the pointer to the message just read becuase
one can figure out the length of the message previously read by invoking
strlen on it (unless of course its NULL, but then we know its 0).

This will be useful down the line.