Add the new AT engine into the codebase
Since we now have a need to be able to handle URCs and AT commands
like a pro, we need an engine that will allow us to do so in a clean
and stateful way.  This is that engine.  It handles everything from
URCs, the command responses, to issuing of commands, even backoffs
after commands to allow for buffered URCs to flow into the device.

The code isn't tested at this point, but it does compile.  More to
come.