Basic WiFi init working
Adds in the esp8266 driver code plus the new wifi task that
drives the whole thing.  Also includes some bug fixes to the
at subsystem to allow for multiple delimeterst to be set
(for "\r\n" in example).  Also fixes a crash bug where I forgot
a couple of `return` statements in a switch.  Added a unit test
to cover that now too.