Fix USB suspend handling
It used to be that we would suspend our USB software stack device when
the USB hardware went into suspend mode and would never re-init it.
The problem with this is that on systems that do aggressive USB sleep,
(Linux 3.4+), this causes the device to never recover once it has gone
to sleep.

This patch addresses that issue by reinitializing the stack when we
see a character (obviously meaning we have come out of resume).  If our
stack isn't awake, it revives it.  This allows us to prevent IRQ storms
when there is no USB, but also recover when it comes back.

Fixes issue #426