Fix Timer pre-scalar skew in MK2 HAL layer
Turns out that we were not correctly adjusting our ticks per
microsecond (us) calculation when the prescaler changed.  This was
resulting in incorrect period readings of 1/2 the speed in fast mode,
and 10x the value in slow mode.  This patch addresses that issue and
consolidates the code a bit more to make it more readable and
understandable.

Fixes issue #435