Improve the setLed method in Lua runtime
Traditionally you did not know which LED you were really setting
when you were setting an LED in the Lua runtime.  All you had were
numbers (and even those could be inconsistent).  This fixes that
by allowing the caller to specify _either_ a number _or_ a string
that represents an LED.  Strings are as follows:

* "LED_ERROR"     - The LED used to indicate an error.
* "LED_GPS"       - The LED used for GPS signaling.
* "LED_LOGGER"    - The LED used to indicate logging is active.
* "LED_TELEMETRY" - The LED used to indicate telemetry is active
                    (when available).

The method will also now return true or false depending on whether
or not the call was successful, with a value of true meaning success.

This fix also re-indexes the LEDs so that they are read left to
right, top to bottom (like a book).  Red LED is 0, the remaining
configurable ones count downward.

This also fixes a corruption bug that could happen if the Lua user
input an index that was outside the bounds of what was supported
on the hardware.