Add in timer quiet period time and edge detection control
This patch adds the ability for the user to specify both the timer
edge detection logic (rising or falling) and to specify the timer
quiet period (either automatic (< 0), disabled (0), or manual (< 0)).
This gets us the flexibility to be as dynamic as possible with this
feature while still being user friendly with the automatic setting
which should work in 99% of cases.