Add support to open sockets and send data
This commit adds the support needed to open an arbitrary socket
and send data as needed.  It requred a tweak to the AT command
engine so that we could support AT commands that didn't complete
when a status message was received (a multi-part command if you
will).  But that ended up being a minor issue.