Separate out BT and Cellular Telem for better performance
Traditionally our code would only stream at the max rate of the slower
of the two telemetry devices.  So if the slow device was cellular, then
BT would send telemetry at the slower rate (10Hz) when it could in-fact
support 50Hz.  Pity.  This fixes that by putting the max rate enforcement
into the communication task so that it can selectively send or drop the
telemetry packet as needed.