Re-tool Lua subsystem, make killable if non-responsive.
This change effectively re-tools the Lua subsystem so that it
can be nuked if the Lua script that it is running becomes
caught in a loop and is unresponsive.  It achieves this by
restructuring how the Lua code is invoked and eliminating the
restarts we used to do when we would run out of memory.  The
rationale behind that is if a user us running out of memory
consistenly we should fail fast and loud so they can fix their
script before going on track.  This change also increases the
Lua task stack size to 8K (2K 4byte entries).

Fixes Issue #403