Remove dependencies on non-existant _sbrk
newlib was looking for this symbol because of the use of malloc,
which was a requirement of vsnprintf and friends.  This is not correct
for RCP family since our malloc methods are handled by FreeRTOS.  This
patch addresses this issue by createing memory.c.  These symbols override
the malloc, free, realloc and calloc methods defined in libc, allowing us
to use our own malloc methods instead of those provided by newlib.  This
in turn allows us to use a single malloc system as well as allows us to
start using stdlib malloc, free, realloc and calloc like a proper program
should instead of having to use portMalloc and portFree etc.