Fix DMA priority and disable interrupts we don't need
DMA was set to a higher interrupt priority (1) than what the FreeRTOS
system was configured for (5). Since we call xQueueSendit from within
that DMA handler and the FreeRTOSConfig.h explicitly states that our
priorities must be >= than the FreeRTOS priority value if the
interrupt calls _any_ FreeRTOS function, we need to reduce it.  Thus I
use GPS interrupt level (6 as of this writing) to handle it.

This change also explicitly disables all interrupt flags to ensure
that we don't get awoken by any silly interrupts we don't care to
handle or deal with.