Re-enable the IMU Task
Was disabled as part of the debug process on RCT.  Now that things
are seeming to be stabilizing again we can add it back.  So far it
has not pee'd the bed as of yet (but I'm literally waiting for it
to do so as I write these very words... waiting... still waiting...
bah... it won't do it until after I save this).  Here is to hoping
it has grown up beyond rubber pants.