Add support for connecting a new socket; abstract Serial onto Sockets
Since we need to open up a connection for our beacon, I created a
_connect method that will open up a new connection.  However internally
we don't yet support Sockets, so I have instead layered a Serial device
ontop of a Socket.  In effect this will behave the same as Serial device
does now: It will drop the data (send to null) if the connection is down,
or it will send it normally if the connection is up.  This is sane for now
but we will need to improve this down the road by introducing Socket support
into our code (see issue #542).  ITMT we can work with this and add a sort
of hacky method to Serial that will allow us to query whether or not we are
working with an open or closed Serial device.  This will allow us to
workaround the shortcomings of a Serial device without paying the full price
of moving to Sockets.

Still need to add a `close` method.  Will do that when necessary (which
probably will be quite soon).